use only 16 bits of timers
even if the timer is 32 bits, use only bottom 16 bits for consistency

+++ Mcu/e230/Inc/blutil.h
@@ -70,7 +70,7 @@ static inline void bl_timer_disable(void)
     timer_disable(BL_TIMER);
 }
 
-static inline uint32_t bl_timer_us(void)
+static inline uint16_t bl_timer_us(void)
 {
     return timer_counter_read(BL_TIMER);
 }
